+++ src/test-integration/kotlin/net/pwall/json/schema/codegen/IntegrationTest.kt
@@ -26,21 +26,68 @@
 package net.pwall.json.schema.codegen
 
 import kotlin.test.Test
+import kotlin.test.assertFailsWith
 import kotlin.test.expect
 
 import java.util.UUID
 
+import net.pwall.json.schema.codegen.test.TestArrayItems
 import net.pwall.json.schema.test.person.Person
 
 class IntegrationTest {
 
     @Test fun `should compile simple generated code`() {
-        println("Running integration tests")
         val uuid = UUID.fromString("c9cfc320-9591-11eb-920e-a78a48655abd")
         val name = "Test Name"
         val person = Person(uuid, name)
         expect(uuid) { person.id }
         expect(name) { person.name }
+    }
+
+    @Test fun `should compile complex code with array validations`() {
+        val testAAA = listOf(123, 456)
+        val testBBB = listOf("AUD", "NZD")
+        val test1 = TestArrayItems(testAAA, testBBB, listOf(testBBB))
+        expect(testAAA) { test1.aaa }
+        expect(testBBB) { test1.bbb }
+        expect(1) { test1.ccc?.size }
+        expect(testBBB) { test1.ccc?.first() }
+        val test2 = TestArrayItems(testAAA, testBBB)
+        expect(testAAA) { test2.aaa }
+        expect(testBBB) { test2.bbb }
+        expect(null) { test2.ccc }
+    }
+
+    @Test fun `should compile complex code with array validations - error cases`() {
+        val testAAA = listOf(123, 456)
+        val testBBB = listOf("AUD", "NZD")
+        val testAAAError1 = listOf(123, 123456)
+        assertFailsWith<IllegalArgumentException> { TestArrayItems(testAAAError1, testBBB, listOf(testBBB)) }.let {
+            expect("aaa item not in range 0..9999 - 123456") { it.message }
+        }
+
+        val testAAAError2 = listOf(123, 456, 789, 987, 765, 432)
+        assertFailsWith<IllegalArgumentException> { TestArrayItems(testAAAError2, testBBB, listOf(testBBB)) }.let {
+            expect("aaa length not in range 1..5 - 6") { it.message }
+        }
+
+        val testBBBError1 = listOf("ABCD", "USD")
+        assertFailsWith<IllegalArgumentException> { TestArrayItems(testAAA, testBBBError1, listOf(testBBB)) }.let {
+            expect("bbb item does not match pattern ^[A-Z]{3}\$ - ABCD") { it.message }
+        }
+
+        assertFailsWith<IllegalArgumentException> { TestArrayItems(testAAA, testBBB, listOf(testBBBError1)) }.let {
+            expect("ccc item item length > maximum 3 - 4") { it.message }
+        }
+
+        val testBBBError2 = listOf("GBP", "USD", "EUR")
+        assertFailsWith<IllegalArgumentException> { TestArrayItems(testAAA, testBBB, listOf(testBBBError2)) }.let {
+            expect("ccc item length not constant value 2 - 3") { it.message }
+        }
+
+        assertFailsWith<IllegalArgumentException> { TestArrayItems(testAAA, testBBB, emptyList()) }.let {
+            expect("ccc length < minimum 1 - 0") { it.message }
+        }
     }
 
 }